Assignments/week08/Olist_Ecommerce/dashboard/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pandas==3.0.1
